third_party/blink/web_tests/svg/dom/SVGGeometryElement-valid-arguments.html

<!DOCTYPE html>
<title>Check invalid arguments of SVGGeometryElement methods</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
  var pathElement = document.createElementNS("http://www.w3.org/2000/svg","path");

  assert_throws_js(TypeError, function() { pathElement.getPointAtLength(); });
  assert_throws_js(TypeError, function() { pathElement.getPointAtLength(NaN); });
  assert_throws_js(TypeError, function() { pathElement.getPointAtLength(Infinity); });
});
</script>
